refactor(codex): split codegraph session hook
