+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/01-pre-edit-loc.txt
@@ -1,0 +1,5 @@
+$ awk '!/^[[:space:]]*$/ && !/^[[:space:]]*(\/\/|#|--)/' packages/omo-codex/plugin/components/codegraph/src/hook.ts | wc -l
+     269
+
+$ wc -l packages/omo-codex/plugin/components/codegraph/src/hook.ts
+     313 packages/omo-codex/plugin/components/codegraph/src/hook.ts
