+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/03-static-scan.txt
@@ -1,0 +1,1 @@
+$ rg -n 'as any|as unknown|@ts-ignore|@ts-expect-error|catch \([^)]*\) \{\s*\}' packages/omo-codex/plugin/components/codegraph/src || true
