+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/04a-component-tests-exact-command.txt
@@ -1,0 +1,12 @@
+$ bun test packages/omo-codex/plugin/components/codegraph/test/hook.test.ts packages/omo-codex/plugin/components/codegraph/test/serve.test.ts
+bun test v1.3.14 (0d9b296a)
+The following filters did not match any test files:
+ packages/omo-codex/plugin/components/codegraph/test/hook.test.ts packages/omo-codex/plugin/components/codegraph/test/serve.test.ts
+6220 files were searched [36.00ms]
+
+note: Tests need ".test", "_test_", ".spec" or "_spec_" in the filename (ex: "MyApp.test.ts")
+note: To treat the "packages/omo-codex/plugin/components/codegraph/test/hook.test.ts" filter as a path, run "bun test ./packages/omo-codex/plugin/components/codegraph/test/hook.test.ts"
+
+Learn more about bun test: https://bun.com/docs/cli/test
+
+[exit] 1
