+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/04d-component-tests-package-script.txt
@@ -1,0 +1,33 @@
+$ bun run --cwd packages/omo-codex/plugin/components/codegraph test
+$ bun test test/*.test.ts
+bun test v1.3.14 (0d9b296a)
+
+test/serve.test.ts:
+(pass) runCodegraphServe > #given CodeGraph is unresolved #when serving MCP #then exits non-zero with a one-line skip hint [1.82ms]
+(pass) runCodegraphServe > #given CodeGraph resolves #when serving MCP #then execs codegraph serve --mcp with inherited stdio and telemetry disabled [0.65ms]
+(pass) runCodegraphServe > #given OMO_CODEGRAPH_BIN points at a missing path #when serving MCP #then exits before spawn [0.20ms]
+(pass) runCodegraphServe > #given Codex SOT disables CodeGraph #when serving MCP #then exits non-zero with a disabled hint [2.24ms]
+(pass) runCodegraphServe > #given Windows Codex SOT install_dir has codegraph.cmd #when serving MCP #then it resolves there and exports CODEGRAPH_INSTALL_DIR [0.97ms]
+(pass) runCodegraphServe > #given built serve entry #when invoked with a fake CodeGraph binary #then it runs serve mcp exactly once [400.47ms]
+(pass) runCodegraphServe > #given built cli entry #when invoked with a fake CodeGraph binary #then it runs serve mcp exactly once [211.08ms]
+
+test/hook.test.ts:
+(pass) CodeGraph SessionStart hook > #given hook session-start cli args #when invoked with empty JSON input #then it emits valid JSON and exits zero [2.09ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is disabled by Codex SOT config #when SessionStart fires #then it skips without spawning [0.13ms]
+(pass) CodeGraph SessionStart hook > #given HOME OMO config disables Codex CodeGraph #when SessionStart fires #then it skips without spawning [0.99ms]
+(pass) CodeGraph SessionStart hook > #given project Codex SOT disables global CodeGraph enablement #when SessionStart fires #then project config wins [0.96ms]
+(pass) CodeGraph SessionStart hook > #given env disables CodeGraph over SOT enablement #when SessionStart fires #then it skips without spawning [0.79ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is enabled #when SessionStart fires #then it detaches a background worker immediately [0.68ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph cannot be resolved or provisioned #when worker runs #then it logs a graceful skip [0.85ms]
+(pass) CodeGraph SessionStart hook > #given CodeGraph is unavailable and auto provisioning is disabled #when worker runs #then it leaves the project untouched [0.63ms]
+(pass) CodeGraph SessionStart hook > #given Windows install_dir has codegraph.cmd #when worker resolves provisioned CodeGraph #then it uses the cmd shim [1.35ms]
+(pass) CodeGraph SessionStart hook > #given resolved CodeGraph status #when worker runs #then it runs status before init or sync [1.37ms]
+(pass) CodeGraph SessionStart hook > #given malformed hook input #when SessionStart fires #then it still emits JSON and exits zero [0.16ms]
+(pass) CodeGraph SessionStart hook > #given plugin hook config #when inspected #then CodeGraph is registered after bootstrap SessionStart [0.23ms]
+
+ 19 pass
+ 0 fail
+ 60 expect() calls
+Ran 19 tests across 2 files. [653.00ms]
+
+[exit] 0
