+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/05-typecheck.txt
@@ -1,0 +1,4 @@
+$ bun run --cwd packages/omo-codex/plugin/components/codegraph typecheck
+$ tsc --noEmit
+
+[exit] 0
