+++ .omo/evidence/20260615-codegraph-omo-integration/final-loc-refactor/06-build.txt
@@ -1,0 +1,12 @@
+$ bun run --cwd packages/omo-codex/plugin/components/codegraph build
+$ bun build src/serve.ts --target node --format esm --outfile dist/serve.js && bun build src/cli.ts --target node --format esm --outfile dist/cli.js
+Bundled 17 modules in 7ms
+
+  serve.js  48.64 KB  (entry point)
+
+Bundled 23 modules in 4ms
+
+  cli.js  73.43 KB  (entry point)
+
+
+[exit] 0
